requirements.txt
crewai
pdf2image
Pillow
requests
python-dotenv
fastapi
uvicorn
pandas
python-multipart
